requirements.txt
ipython>=9.2.0
django-extensions>=3.2.1
djangorestframework>=3.14.0
django-cors-headers>=3.14.0
